To view the content you need to open the program "Mini programs"
Content

1. All programs
2. Their functions
3. Explanations of the principle of their work
4. errors

There are three programs at the moment: Mini guess the number game, Random graphic password generator and Calculator.

In the mini-game, you choose the difficulty; the range of the guessed number and the number of attempts depend on it. You write what number you think was made up by the computer. If you have not guessed, one attempt is taken away from you, but a hint is given, your number is more or less in comparison with the number of the computer. After winning or losing, you have the opportunity to play the game or return to the list of programs.

In the random graphic password generator, you can set the password length from 5 - 9. The password is indicated by the order of numbers to be followed. It is possible to generate another password and return to the list of programs.

In the calculator, you can enter three numbers and specify two of the 6 possible actions

The mini-game generates a random number in the range of those numbers which set difficulty levels. Then he lets the player enter his number. The program compares two numbers and determines whether the player has guessed or not. If not, the program takes one attempt from the player and gives a hint by comparing the number of the player and the computer.

the random number generator generates the first number and the second then compares whether they coincide; if so, the computer generates this number if the values ​​match again; the computer repeats the procedure. after that it generates a third number and compares it with the others; if it coincides with one of them, it again regenerates the number. So everything lasts until the fifth day. Later, the computer checks what length the user has set and whether it is necessary to generate the sixth number and so on with all the others. At the end, the computer displays the resulting password.